pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terminalbrain"
version = "0.1.0"
description = "AI-powered terminal assistant and command recommendation engine"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Terminal Brain Team"}]
keywords = ["terminal", "ai", "cli", "command-prediction", "ml"]

dependencies = [
    "textual>=0.38.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
    "pyyaml>=6.0",
    "toml>=0.10.0",
    "ollama>=0.0.11",
    "openai>=1.0.0",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.0",
    "psutil>=5.9.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
gpu = [
    "faiss-gpu>=1.7.4",
    "torch>=2.0.0",
]

[project.scripts]
terminalbrain = "terminalbrain.cli:main"
tb = "terminalbrain.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["terminalbrain*"]

[tool.setuptools.package-data]
terminalbrain = ["workflows/*.yaml", "shell/*.sh"]
